Refactor user routes to new file
